+++ src/api/invoices.rs
@@ -1,5 +1,3 @@
-use crate::api::utils::is_valid_iban;
-use crate::api::utils::is_valid_phone_number;
 use std::sync::LazyLock;
 
 use crate::error::Error;
@@ -10,10 +8,10 @@ use axum_typed_multipart::{
     FieldData, FieldMetadata, TryFromChunks, TryFromMultipart, TypedMultipart, TypedMultipartError,
 };
 
-use crate::api::utils;
 use axum_valid::Garde;
 use futures::stream::Stream;
 use garde::Validate;
+use iban::Iban;
 use regex::Regex;
 use serde_derive::{Deserialize, Serialize};
 use utoipa::ToSchema;
@@ -27,7 +25,33 @@ impl TryFromChunks for Invoice {
         chunks: impl Stream<Item = Result<Bytes, TypedMultipartError>> + Send + Sync + Unpin,
         metadata: FieldMetadata,
     ) -> Result<Self, TypedMultipartError> {
-        utils::handle_chunks(chunks, metadata).await
+        let bytes = Bytes::try_from_chunks(chunks, metadata).await?;
+
+        serde_json::from_slice(&bytes).map_err(|e| TypedMultipartError::Other { source: e.into() })
+    }
+}
+
+pub fn is_valid_iban(value: &str, _: &()) -> garde::Result {
+    match value.parse::<Iban>() {
+        Err(e) => Err(garde::Error::new(e)),
+        _ => Ok(()),
+    }
+}
+
+pub fn is_valid_phone_number(value: &str, _: &()) -> garde::Result {
+    use phonenumber::country::Id;
+    // Works if number is in international format
+    if phonenumber::parse(None, value)
+        .map(|number| number.is_valid())
+        .unwrap_or(false)
+    {
+        return Ok(());
+    }
+    // Missing country code but number is otherwise valid, assume FI
+    match phonenumber::parse(Some(Id::FI), value).map(|n| n.is_valid()) {
+        Ok(true) => Ok(()),
+        Err(e) => Err(garde::Error::new(format!("not a valid phone number: {e}"))),
+        _ => Err(garde::Error::new("Invalid phone number")),
     }
 }
