+++ src/api/mod.rs
@@ -17,7 +17,6 @@ use crate::{api::key_extractor::IpExtractor, CONFIG};
 pub mod invoices;
 mod key_extractor;
 pub mod receipts;
-pub mod utils;
 
 pub fn app() -> Router<crate::state::State> {
     let cors_layer = CorsLayer::new().allow_origin(
